Update Finance_Hub.java
I have added two currency options to the program which will enhance the program's features. I am looking forward to change the interface to make it more accessible through web platforms but at the moment I will stick with this first. Please review the work and check if the code I applied is appropriate.